refactor: Unify json serialization for saga module
