+++ compatible/src/test/java/io/seata/saga/engine/StateMachineTests.java
@@ -22,7 +22,7 @@ import io.seata.saga.engine.mock.DemoService.People;
 import io.seata.saga.statelang.domain.DomainConstants;
 import io.seata.saga.statelang.domain.ExecutionStatus;
 import io.seata.saga.statelang.domain.StateMachineInstance;
-import org.apache.seata.saga.statelang.parser.JsonParserFactory;
+import org.apache.seata.json.JsonParserFactory;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
@@ -263,7 +263,7 @@ public class StateMachineTests {
         map1.put("lilei4", people4);
         people1.setChildrenMap(map1);
 
-        String json = JsonParserFactory.getJsonParser("jackson").toJsonString(people1, false, true);
+        String json = JsonParserFactory.getInstance().toJsonString(people1, false, true);
         System.out.println(json);
     }
