+++ pom.xml
@@ -72,6 +72,7 @@
         <module>ext/apm-seata-skywalking-plugin</module>
         <module>integration-tx-api</module>
         <module>namingserver</module>
+        <module>json</module>
         <!--<module>seata-plugin</module>-->
     </modules>
