+++ saga/seata-saga-spring/src/main/java/org/apache/seata/saga/engine/config/DbStateMachineConfig.java
@@ -28,7 +28,6 @@ import org.apache.seata.saga.engine.tm.SagaTransactionalTemplate;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.DisposableBean;
-import org.springframework.util.StringUtils;
 
 import javax.sql.DataSource;
 import java.sql.Connection;
@@ -39,7 +38,6 @@ import static org.apache.seata.common.DefaultValues.DEFAULT_CLIENT_REPORT_SUCCES
 import static org.apache.seata.common.DefaultValues.DEFAULT_CLIENT_SAGA_BRANCH_REGISTER_ENABLE;
 import static org.apache.seata.common.DefaultValues.DEFAULT_CLIENT_SAGA_COMPENSATE_PERSIST_MODE_UPDATE;
 import static org.apache.seata.common.DefaultValues.DEFAULT_CLIENT_SAGA_RETRY_PERSIST_MODE_UPDATE;
-import static org.apache.seata.common.DefaultValues.DEFAULT_SAGA_JSON_PARSER;
 
 /**
  * DbStateMachineConfig
@@ -66,8 +64,6 @@ public class DbStateMachineConfig extends DefaultStateMachineConfig implements D
                 setSagaBranchRegisterEnable(configuration.getBoolean(
                         ConfigurationKeys.CLIENT_SAGA_BRANCH_REGISTER_ENABLE,
                         DEFAULT_CLIENT_SAGA_BRANCH_REGISTER_ENABLE));
-                setSagaJsonParser(
-                        configuration.getConfig(ConfigurationKeys.CLIENT_SAGA_JSON_PARSER, DEFAULT_SAGA_JSON_PARSER));
                 this.applicationId = configuration.getConfig(ConfigurationKeys.APPLICATION_ID);
                 this.txServiceGroup = configuration.getConfig(ConfigurationKeys.TX_SERVICE_GROUP);
                 this.accessKey = configuration.getConfig(ConfigurationKeys.ACCESS_KEY, null);
@@ -105,12 +101,7 @@ public class DbStateMachineConfig extends DefaultStateMachineConfig implements D
             dbStateLogStore.setDbType(dbType);
             dbStateLogStore.setDefaultTenantId(getDefaultTenantId());
             dbStateLogStore.setSeqGenerator(getSeqGenerator());
-
-            if (StringUtils.hasLength(getSagaJsonParser())) {
-                ParamsSerializer paramsSerializer = new ParamsSerializer();
-                paramsSerializer.setJsonParserName(getSagaJsonParser());
-                dbStateLogStore.setParamsSerializer(paramsSerializer);
-            }
+            dbStateLogStore.setParamsSerializer(new ParamsSerializer());
 
             if (sagaTransactionalTemplate == null) {
                 DefaultSagaTransactionalTemplate defaultSagaTransactionalTemplate =
