+++ saga/seata-saga-statelang/src/main/java/org/apache/seata/saga/statelang/parser/StateMachineParserFactory.java
@@ -24,7 +24,7 @@ import org.apache.seata.saga.statelang.parser.impl.StateMachineParserImpl;
  */
 public class StateMachineParserFactory {
 
-    public static StateMachineParser getStateMachineParser(String jsonParserName) {
-        return new StateMachineParserImpl(jsonParserName);
+    public static StateMachineParser getStateMachineParser() {
+        return new StateMachineParserImpl();
     }
 }
