+++ saga/seata-saga-statelang/src/test/resources/file.conf
@@ -14,5 +14,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-org.apache.seata.saga.statelang.parser.impl.FastjsonParser
-org.apache.seata.saga.statelang.parser.impl.JacksonJsonParser
+
+jsonParser {
+  type = "jackson"
+}
