+++ saga/seata-saga-statelang/src/test/resources/registry.conf
@@ -14,23 +14,16 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-
-service {
-  #transaction service group mapping
-  vgroupMapping.default_tx_group = "default"
-  #only support when registry.type=file, please don't set multiple addresses
-  default.grouplist = "127.0.0.1:8091"
-  #disable seata
-  disableGlobalTransaction = false
-}
-
-client {
-  rm {
-    sagaRetryPersistModeUpdate = false
-    sagaCompensatePersistModeUpdate = false
+registry {
+  type = "file"
+  file {
+    name = "file.conf"
   }
 }
 
-jsonParser {
-  type = "jackson"
+config {
+  type = "file"
+  file {
+    name = "file.conf"
+  }
 }
