+++ script/client/conf/file.conf
@@ -79,7 +79,6 @@ client {
     tableMetaCheckerInterval = 60000
     reportSuccessEnable = false
     sagaBranchRegisterEnable = false
-    sagaJsonParser = "fastjson"
     sagaRetryPersistModeUpdate = false
     sagaCompensatePersistModeUpdate = false
     tccActionInterceptorOrder = -2147482648 #Ordered.HIGHEST_PRECEDENCE + 1000
@@ -127,4 +126,7 @@ tcc {
     # tcc fence log clean period
     cleanPeriod = 1h
   }
+}
+jsonParser {
+  type = "fastjson"
 }
