+++ test-old-version/src/test/resources/file.conf
@@ -67,7 +67,6 @@ client {
   rm {
     reportSuccessEnable = false
     sagaBranchRegisterEnable = false
-    sagaJsonParser = jackson
     sagaRetryPersistModeUpdate = false
     sagaCompensatePersistModeUpdate = false
   }
@@ -76,3 +75,8 @@ client {
       virtualNodes = 10
   }
 }
+
+jsonParser {
+  type = "jackson"
+}
+
