+++ test/src/test/java/org/apache/seata/saga/engine/StateMachineTests.java
@@ -17,12 +17,12 @@
 package org.apache.seata.saga.engine;
 
 import org.apache.seata.common.SagaCostPrint;
+import org.apache.seata.json.JsonParserFactory;
 import org.apache.seata.saga.engine.mock.DemoService.Engineer;
 import org.apache.seata.saga.engine.mock.DemoService.People;
 import org.apache.seata.saga.statelang.domain.DomainConstants;
 import org.apache.seata.saga.statelang.domain.ExecutionStatus;
 import org.apache.seata.saga.statelang.domain.StateMachineInstance;
-import org.apache.seata.saga.statelang.parser.JsonParserFactory;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
@@ -263,7 +263,7 @@ public class StateMachineTests {
         map1.put("lilei4", people4);
         people1.setChildrenMap(map1);
 
-        String json = JsonParserFactory.getJsonParser("jackson").toJsonString(people1, false, true);
+        String json = JsonParserFactory.getInstance().toJsonString(people1, false, true);
         System.out.println(json);
     }
