+++ test/src/test/resources/file.conf
@@ -67,7 +67,6 @@ client {
   rm {
     reportSuccessEnable = false
     sagaBranchRegisterEnable = false
-    sagaJsonParser = jackson
     sagaRetryPersistModeUpdate = false
     sagaCompensatePersistModeUpdate = false
   }
@@ -75,4 +74,8 @@ client {
       type = "XID"
       virtualNodes = 10
   }
+}
+
+jsonParser {
+  type = "jackson"
 }
